use tabs in op settings
setting/zone/perm
