--- src/code/dialogs/opPerms.d.ts
@@ -1,4 +1,0 @@
-import { WDialog } from "../leafletClasses";
-
-declare class OpPermList extends WDialog {}
-export default OpPermList;
