+++ src/code/dialogs/zoneDrawHandler.d.ts
@@ -1,0 +1,9 @@
+import { WDialog } from "../leafletClasses";
+
+interface ZonedrawHandlerOptions {
+  parent: WDialog;
+}
+export class ZonedrawHandler extends L.Handler {
+  zoneID: number;
+  constructor(map: L.Map, options: ZonedrawHandlerOptions);
+}
